+++ example/app/AppSync.tsx
@@ -1,4 +1,6 @@
 import React, {useEffect, useMemo} from 'react';
+import {useApp} from '@realm/react';
+import {StyleSheet, Text} from 'react-native';
 
 import {Task} from './models/Task';
 import {TaskRealmContext} from './models';
@@ -12,6 +14,7 @@ type AppProps = {
 
 export const AppSync: React.FC<AppProps> = ({userId}) => {
   const realm = useRealm();
+  const app = useApp();
   const result = useQuery(Task);
 
   const tasks = useMemo(() => result.sorted('createdAt'), [result]);
@@ -22,5 +25,17 @@ export const AppSync: React.FC<AppProps> = ({userId}) => {
     });
   }, [realm, result]);
 
-  return <TaskManager tasks={tasks} userId={userId} />;
+  return (
+    <>
+      <Text style={styles.idText}>Syncing with app id: {app.id}</Text>
+      <TaskManager tasks={tasks} userId={userId} />
+    </>
+  );
 };
+
+const styles = StyleSheet.create({
+  idText: {
+    color: '#999',
+    paddingHorizontal: 20,
+  },
+});